+++ lib/routes/monetization_bottom_banner_ad_ex.dart
@@ -1,8 +1,7 @@
-import 'dart:io';
-
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
+
+import 'feature_store_secrets.dart';
 
 class BottomBannerAdExample extends StatefulWidget {
   const BottomBannerAdExample({super.key});
@@ -15,22 +14,6 @@ class _BottomBannerAdExampleState extends State<BottomBannerAdExample> {
   bool _adLoaded = false;
   late BannerAd _bannerAd;
 
-  String get _kAdUnitId {
-    // ! Return test ad unit if we are in debug mode -- otherwise account might be banned!
-    if (Platform.isAndroid) {
-      return kDebugMode
-          // https://developers.google.com/admob/android/test-ads#sample_ad_units
-          ? 'ca-app-pub-3940256099942544/6300978111'
-          : 'ca-app-pub-7906158617398863/4414540775';
-    } else if (Platform.isIOS) {
-      return kDebugMode
-          // https://developers.google.com/admob/ios/test-ads#demo_ad_units
-          ? 'ca-app-pub-3940256099942544/2934735716'
-          : 'ca-app-pub-7906158617398863/5234412829';
-    }
-    return '';
-  }
-
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
@@ -39,7 +22,7 @@ class _BottomBannerAdExampleState extends State<BottomBannerAdExample> {
 
   Future<void> _loadAd() async {
     _bannerAd = BannerAd(
-      adUnitId: _kAdUnitId,
+      adUnitId: MySecretsHelper.bannerAdUnitId,
       size: AdSize.banner,
       request: AdRequest(),
       listener: BannerAdListener(